implements alternateColors() to glow squares
This commit completes the Glow.pde experiment in which I was trying to
figure out how to reliably glow between two colors.  I have also
implemented the function into the Chess program so that a square you
select will now glow between its default color and a glow_color that is
set in the Game class.